requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
aiosmtplib==3.0.1
google-api-python-client==2.193.0
google-auth-oauthlib==1.3.0
google-auth-httplib2==0.3.0
email-validator==2.2.0
aiofiles==24.1.0
